chore: move "web3Buttons" to sharedComponents folder
