+++ src/pageComponents/home/Examples/demos/ERC20ApproveAndTransferButtonDemo.tsx
@@ -5,8 +5,8 @@ import { useWriteContract } from 'wagmi'
 
 import { useSuspenseReadErc20BalanceOf } from '@/src/hooks/generated'
 import { useWeb3StatusConnected } from '@/src/hooks/useWeb3Status'
+import ERC20ApproveAndTransferButton from '@/src/sharedComponents/ERC20ApproveAndTransferButton'
 import { withWalletStatusVerifier } from '@/src/sharedComponents/WalletStatusVerifier'
-import ERC20ApproveAndTransferButton from '@/src/sharedComponents/Web3Buttons/ERC20ApproveAndTransferButton'
 import { type Token } from '@/src/types/token'
 import { formatNumberOrString, NumberType } from '@/src/utils/format'
 import { withSuspense } from '@/src/utils/suspenseWrapper'
